build: constrain Tutor versions in base.in
Theoretically, the constraint should be:

  tutor>=14,<15

because this plugin targets Nutmeg (ie Tutor v14).
Since Nutmeg isn't out yet, though, Tutor v14 isn't released,
so the "correct" constraint would be impossible to fulfill.
In the mean time, we declare this requirement as:

  tutor>=13.1,<15

Upon Tutor v14's, release, we'll change it to:

  tutor>=14,<15